Prevent README.pod being installed to @INC
Due to historical reasons, anything ending with ".pod" or ".pm"
in the distribution root directory is treated as installable pm or .pod
files, and are installed to the host @INC.

Hence, "README.pod" turns up on the host as:

 lib/HTML/FormatText/WithLinks/README.pod

Because the last path-part of the distribution is removed and used as
a basename.

INSTALL.SKIP directs ExtUtils::Install not to copy any files matching
any regular expressions in that file from blib/ to @INC

Alternative solutions involve using only README.md in the shipped
distribution instead.